refactor: separate markdown parsing and rendering
