+++ app/src/main/java/com/echoflow/ui/components/MarkdownCodeBlock.kt
@@ -1,0 +1,162 @@
+
+package com.echoflow.ui.components
+
+import android.content.ClipData
+import android.content.ClipboardManager
+import android.content.Context
+import android.widget.Toast
+import androidx.compose.foundation.background
+import androidx.compose.foundation.horizontalScroll
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.InlineTextContent
+import androidx.compose.foundation.text.appendInlineContent
+import androidx.compose.foundation.text.selection.SelectionContainer
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ContentCopy
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.text.AnnotatedString
+import androidx.compose.ui.text.LinkAnnotation
+import androidx.compose.ui.text.Placeholder
+import androidx.compose.ui.text.PlaceholderVerticalAlign
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.TextLinkStyles
+import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.BaselineShift
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextDecoration
+import androidx.compose.ui.text.withLink
+import androidx.compose.ui.text.withStyle
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.unit.TextUnit
+import dev.snipme.highlights.Highlights
+import dev.snipme.highlights.model.BoldHighlight
+import dev.snipme.highlights.model.ColorHighlight
+import dev.snipme.highlights.model.SyntaxLanguage
+import dev.snipme.highlights.model.SyntaxThemes
+import com.hrm.latex.renderer.Latex
+import com.hrm.latex.renderer.LatexAutoWrap
+import com.hrm.latex.renderer.measure.rememberLatexMeasurer
+import com.hrm.latex.renderer.model.LatexConfig
+import com.hrm.latex.renderer.model.LatexTheme
+
+/**
+ * One markdown renderer for the whole app — used identically while a message streams and once it is
+ * finished, so the layout never "jumps" or re-styles when streaming completes. It is a lightweight,
+ * block-incremental CommonMark/GFM renderer: parsing is memoized on the text and every block is a
+ * child composable taking only stable params, so during streaming only the final, still-growing
+ * block re-renders each frame. Supports headers, ordered/bulleted (nested) lists, blockquotes,
+ * horizontal rules, GFM tables (wrapping cells — never truncated), fenced code blocks with real
+ * syntax highlighting, LaTeX math (`$...$`, `$$...$$`, `\(...\)`, `\[...\]`), and inline
+ * bold/italic/strikethrough/code/links.
+ */
+
+@Composable
+fun CodeBlockItem(code: String, language: String?) {
+    val context = LocalContext.current
+    val darkTheme = isSystemInDarkTheme()
+    // Real syntax highlighting via the Highlights engine, memoized per (code, language, theme) so
+    // finished blocks never recompute and only the actively-streaming block re-tokenizes.
+    val highlighted = remember(code, language, darkTheme) {
+        highlightCode(code, language, darkTheme)
+    }
+    Surface(
+        modifier = Modifier.fillMaxWidth().padding(vertical = 4.dp),
+        shape = RoundedCornerShape(16.dp),
+        color = MaterialTheme.colorScheme.surfaceContainerHighest,
+    ) {
+        Column {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .background(MaterialTheme.colorScheme.surfaceContainerHigh)
+                    .padding(start = 16.dp, end = 8.dp, top = 6.dp, bottom = 6.dp),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text(
+                    text = (language ?: "code").uppercase(),
+                    style = MaterialTheme.typography.labelMedium.copy(
+                        fontFamily = FontFamily.Monospace,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                )
+                IconButton(
+                    onClick = {
+                        val clipboard = context.getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
+                        val clip = ClipData.newPlainText("Copied Code", code)
+                        clipboard.setPrimaryClip(clip)
+                        Toast.makeText(context, "Code copied into clipboard", Toast.LENGTH_SHORT).show()
+                    },
+                    modifier = Modifier.size(28.dp)
+                ) {
+                    Icon(
+                        imageVector = Icons.Default.ContentCopy,
+                        contentDescription = "Copy code",
+                        tint = MaterialTheme.colorScheme.onSurfaceVariant,
+                        modifier = Modifier.size(16.dp)
+                    )
+                }
+            }
+            Box(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .horizontalScroll(rememberScrollState())
+                    .padding(12.dp)
+            ) {
+                SelectionContainer {
+                    Text(
+                        text = highlighted,
+                        style = MaterialTheme.typography.bodyMedium.copy(
+                            color = MaterialTheme.colorScheme.onSurface,
+                            fontFamily = FontFamily.Monospace,
+                            fontSize = 13.sp,
+                            lineHeight = 18.sp
+                        )
+                    )
+                }
+            }
+        }
+    }
+}
+
+/** Tokenizes [code] with the Highlights engine and maps the result to a styled [AnnotatedString]. */
+internal fun highlightCode(code: String, language: String?, darkTheme: Boolean): AnnotatedString {
+    val syntaxLanguage = language?.let { runCatching { SyntaxLanguage.getByName(it) }.getOrNull() }
+    val highlights = runCatching {
+        Highlights.Builder()
+            .code(code)
+            .theme(SyntaxThemes.default(darkMode = darkTheme))
+            .let { if (syntaxLanguage != null) it.language(syntaxLanguage) else it }
+            .build()
+            .getHighlights()
+    }.getOrNull().orEmpty()
+
+    return buildAnnotatedString {
+        append(code)
+        highlights.forEach { h ->
+            val start = h.location.start.coerceIn(0, code.length)
+            val end = h.location.end.coerceIn(start, code.length)
+            if (start == end) return@forEach
+            when (h) {
+                is ColorHighlight -> addStyle(SpanStyle(color = Color(h.rgb).copy(alpha = 1f)), start, end)
+                is BoldHighlight -> addStyle(SpanStyle(fontWeight = FontWeight.Bold), start, end)
+            }
+        }
+    }
+}
+
