+++ app/src/main/java/com/echoflow/ui/components/MarkdownText.kt
@@ -1,3 +1,4 @@
+
 package com.echoflow.ui.components
 
 import android.content.ClipData
@@ -245,809 +246,6 @@ private fun TableRow(
                     .weight(1f)
                     .padding(horizontal = 10.dp, vertical = 8.dp),
             )
-        }
-    }
-}
-
-sealed class MarkdownBlock {
-    data class Header(val text: String, val level: Int) : MarkdownBlock()
-    data class CodeBlock(val code: String, val language: String?) : MarkdownBlock()
-    data class MathBlock(val latex: String, val raw: String, val complete: Boolean) : MarkdownBlock()
-    /** A list item. [ordinal] is null for bullets, the number for ordered items. */
-    data class BulletItem(val text: String, val indent: Int, val ordinal: Int?) : MarkdownBlock()
-    data class Quote(val text: String) : MarkdownBlock()
-    object Divider : MarkdownBlock()
-    data class Table(val headers: List<String>, val aligns: List<TextAlign>, val rows: List<List<String>>) : MarkdownBlock()
-    data class Paragraph(val text: String) : MarkdownBlock()
-}
-
-private val ORDERED_ITEM = Regex("""^(\d+)[.)]\s+(.*)""")
-private val BULLET_ITEM = Regex("""^[*\-+]\s+(.*)""")
-private val HR_LINE = Regex("""^([-*_])\1{2,}\s*$""")
-private val TABLE_SEPARATOR = Regex("""^\s*\|?\s*:?-{1,}:?\s*(\|\s*:?-{1,}:?\s*)+\|?\s*$""")
-
-/**
- * Parses markdown text into a flat list of renderable blocks. Line-based and allocation-light so it
- * is cheap to re-run on every streaming frame.
- */
-fun parseMarkdownBlocks(text: String): List<MarkdownBlock> {
-    val blocks = mutableListOf<MarkdownBlock>()
-    val lines = text.split("\n")
-
-    var insideCodeBlock = false
-    val codeAccumulator = StringBuilder()
-    var codeLanguage: String? = null
-
-    val paragraph = StringBuilder()
-    fun flushParagraph() {
-        if (paragraph.isNotEmpty()) {
-            blocks.add(MarkdownBlock.Paragraph(paragraph.toString()))
-            paragraph.setLength(0)
-        }
-    }
-
-    var i = 0
-    while (i < lines.size) {
-        val line = lines[i]
-        val trimmed = line.trim()
-
-        // Fenced code blocks (``` or ~~~) — take precedence over everything.
-        if (trimmed.startsWith("```") || trimmed.startsWith("~~~")) {
-            if (insideCodeBlock) {
-                blocks.add(MarkdownBlock.CodeBlock(codeAccumulator.toString().trimEnd('\n'), codeLanguage))
-                codeAccumulator.setLength(0)
-                codeLanguage = null
-                insideCodeBlock = false
-            } else {
-                flushParagraph()
-                insideCodeBlock = true
-                codeLanguage = trimmed.substring(3).trim().substringBefore(' ').ifEmpty { null }
-            }
-            i++
-            continue
-        }
-        if (insideCodeBlock) {
-            codeAccumulator.append(line).append('\n')
-            i++
-            continue
-        }
-
-        parseDisplayMathBlock(lines, i)?.let { parsed ->
-            flushParagraph()
-            blocks.add(parsed.block)
-            i = parsed.nextIndex
-            continue
-        }
-
-        // GFM table: a header line containing '|' immediately followed by a separator row.
-        if (line.contains('|') && i + 1 < lines.size && TABLE_SEPARATOR.matches(lines[i + 1])) {
-            flushParagraph()
-            val headers = splitTableRow(line)
-            val aligns = parseTableAligns(lines[i + 1])
-            val rows = mutableListOf<List<String>>()
-            var j = i + 2
-            while (j < lines.size && lines[j].contains('|') && lines[j].isNotBlank()) {
-                rows.add(splitTableRow(lines[j]))
-                j++
-            }
-            blocks.add(MarkdownBlock.Table(headers, aligns, rows))
-            i = j
-            continue
-        }
-
-        when {
-            trimmed.isEmpty() -> flushParagraph()
-
-            HR_LINE.matches(trimmed) -> {
-                flushParagraph()
-                blocks.add(MarkdownBlock.Divider)
-            }
-
-            trimmed.startsWith("#") -> {
-                val level = trimmed.takeWhile { it == '#' }.length
-                if (level in 1..6 && trimmed.length > level && trimmed[level] == ' ') {
-                    flushParagraph()
-                    blocks.add(MarkdownBlock.Header(trimmed.substring(level + 1).trim(), level))
-                } else {
-                    appendParagraphLine(paragraph, line)
-                }
-            }
-
-            trimmed.startsWith(">") -> {
-                flushParagraph()
-                blocks.add(MarkdownBlock.Quote(trimmed.removePrefix(">").trim()))
-            }
-
-            ORDERED_ITEM.matches(trimmed) -> {
-                flushParagraph()
-                val m = ORDERED_ITEM.find(trimmed)!!
-                blocks.add(MarkdownBlock.BulletItem(m.groupValues[2].trim(), indentOf(line), m.groupValues[1].toIntOrNull() ?: 1))
-            }
-
-            BULLET_ITEM.matches(trimmed) -> {
-                flushParagraph()
-                val m = BULLET_ITEM.find(trimmed)!!
-                blocks.add(MarkdownBlock.BulletItem(m.groupValues[1].trim(), indentOf(line), null))
-            }
-
-            else -> appendParagraphLine(paragraph, line)
-        }
-        i++
-    }
-
-    // Flush trailing state (handles still-streaming, unclosed blocks gracefully).
-    if (insideCodeBlock && codeAccumulator.isNotEmpty()) {
-        blocks.add(MarkdownBlock.CodeBlock(codeAccumulator.toString().trimEnd('\n'), codeLanguage))
-    }
-    flushParagraph()
-    return blocks
-}
-
-/** Per CommonMark a single newline inside a paragraph is a soft break (rendered as a space). */
-private fun appendParagraphLine(sb: StringBuilder, line: String) {
-    if (sb.isNotEmpty()) sb.append(' ')
-    sb.append(line.trim())
-}
-
-private fun indentOf(line: String): Int = (line.takeWhile { it == ' ' }.length / 2).coerceAtMost(4)
-
-private fun splitTableRow(line: String): List<String> =
-    line.trim().removePrefix("|").removeSuffix("|").split("|").map { it.trim() }
-
-private fun parseTableAligns(separator: String): List<TextAlign> =
-    separator.trim().removePrefix("|").removeSuffix("|").split("|").map {
-        val s = it.trim()
-        val left = s.startsWith(":")
-        val right = s.endsWith(":")
-        when {
-            left && right -> TextAlign.Center
-            right -> TextAlign.End
-            else -> TextAlign.Start
-        }
-    }
-
-private data class ParsedDisplayMath(val block: MarkdownBlock.MathBlock, val nextIndex: Int)
-
-private fun parseDisplayMathBlock(lines: List<String>, index: Int): ParsedDisplayMath? {
-    val line = lines[index]
-    val trimmed = line.trim()
-    val delimiter = when {
-        trimmed.startsWith("$$") -> "$$"
-        trimmed.startsWith("\\[") -> "\\]"
-        else -> return null
-    }
-    val open = if (delimiter == "$$") "$$" else "\\["
-    val close = delimiter
-    val firstContent = trimmed.removePrefix(open)
-
-    if (firstContent.contains(close)) {
-        val latex = firstContent.substringBefore(close).trim()
-        val tail = firstContent.substringAfter(close)
-        if (tail.isBlank()) {
-            return ParsedDisplayMath(
-                MarkdownBlock.MathBlock(latex = latex, raw = line, complete = latex.isNotBlank()),
-                index + 1
-            )
-        }
-        return null
-    }
-
-    val raw = StringBuilder(line)
-    val latex = StringBuilder(firstContent.trimStart())
-    var i = index + 1
-    while (i < lines.size) {
-        val current = lines[i]
-        raw.append('\n').append(current)
-        val closeAt = current.indexOf(close)
-        if (closeAt != -1) {
-            if (latex.isNotEmpty()) latex.append('\n')
-            latex.append(current.substring(0, closeAt).trimEnd())
-            return ParsedDisplayMath(
-                MarkdownBlock.MathBlock(
-                    latex = latex.toString().trim(),
-                    raw = raw.toString(),
-                    complete = true
-                ),
-                i + 1
-            )
-        }
-        if (latex.isNotEmpty()) latex.append('\n')
-        latex.append(current)
-        i++
-    }
-
-    return ParsedDisplayMath(
-        MarkdownBlock.MathBlock(
-            latex = latex.toString().trim(),
-            raw = raw.toString(),
-            complete = false
-        ),
-        lines.size
-    )
-}
-
-private fun findInlineDollarClose(text: String, from: Int): Int {
-    var i = from
-    while (i < text.length) {
-        if (text[i] == '$' && !isEscaped(text, i) && text.getOrNull(i + 1) != '$') return i
-        i++
-    }
-    return -1
-}
-
-private fun isEscaped(text: String, index: Int): Boolean {
-    var slashCount = 0
-    var i = index - 1
-    while (i >= 0 && text[i] == '\\') {
-        slashCount++
-        i--
-    }
-    return slashCount % 2 == 1
-}
-
-private fun looksLikeMath(content: String, before: Char?, after: Char?): Boolean {
-    if (content.isBlank()) return false
-    if (content.first().isWhitespace() || content.last().isWhitespace()) return false
-    if (content.length > 160) return false
-    if (before?.isLetterOrDigit() == true && after?.isLetterOrDigit() == true) return false
-    if (content.all { it.isDigit() || it == '.' || it == ',' }) {
-        return before != '$' && after?.isLetter() != true && content.any { it.isDigit() }
-    }
-    val lower = content.lowercase()
-    val hasMathSignal = content.any { it in "\\^_=+-*/()[]{}<>|'" } ||
-        listOf("frac", "sqrt", "lim", "sum", "int", "text", "sin", "cos", "tan", "log", "ln").any { it in lower }
-    val isTinyVariable = content.length <= 4 && content.any { it.isLetter() } && content.none { it.isDigit() }
-    return hasMathSignal || isTinyVariable
-}
-
-private sealed class InlineNode {
-    data class Text(val text: String) : InlineNode()
-    data class Math(val id: Int, val latex: String, val raw: String) : InlineNode()
-    data class Span(val style: SpanStyle, val children: List<InlineNode>) : InlineNode()
-    data class Code(val text: String) : InlineNode()
-    data class Link(val label: List<InlineNode>, val url: String) : InlineNode()
-    data class Citation(val label: String, val url: String) : InlineNode()
-}
-
-private class InlineParser(private val source: String) {
-    private var nextMathId = 0
-
-    fun parse(): List<InlineNode> = parseRange(0, source.length)
-
-    private fun parseRange(start: Int, end: Int): List<InlineNode> {
-        val nodes = mutableListOf<InlineNode>()
-        var cursor = start
-        while (cursor < end) {
-            var nextIdx = -1
-            var kind = ""
-            fun consider(idx: Int, k: String) {
-                if (idx in cursor until end && (nextIdx == -1 || idx < nextIdx)) {
-                    nextIdx = idx
-                    kind = k
-                }
-            }
-            consider(source.indexOf("\\(", cursor).takeIf { it < end } ?: -1, "parenMath")
-            consider(source.indexOf("\\[", cursor).takeIf { it < end } ?: -1, "bracketMath")
-            consider(indexOfDoubleDollar(cursor, end), "doubleDollarMath")
-            consider(indexOfDollar(cursor, end), "dollarMath")
-            consider(source.indexOf("**", cursor).takeIf { it < end } ?: -1, "bold")
-            consider(source.indexOf("~~", cursor).takeIf { it < end } ?: -1, "strike")
-            consider(source.indexOf("`", cursor).takeIf { it < end } ?: -1, "code")
-            consider(source.indexOf("[", cursor).takeIf { it < end } ?: -1, "link")
-            consider(indexOfItalic(source, cursor).takeIf { it < end } ?: -1, "italic")
-
-            if (nextIdx == -1) {
-                nodes.add(InlineNode.Text(source.substring(cursor, end)))
-                break
-            }
-            if (nextIdx > cursor) nodes.add(InlineNode.Text(source.substring(cursor, nextIdx)))
-
-            when (kind) {
-                "parenMath" -> {
-                    val close = source.indexOf("\\)", nextIdx + 2).takeIf { it in 0 until end }
-                    val parsed = close?.let { mathNode(nextIdx + 2, it, it + 2) }
-                    if (parsed != null) {
-                        nodes.add(parsed)
-                        cursor = close + 2
-                    } else {
-                        nodes.add(InlineNode.Text(source.substring(nextIdx, end)))
-                        break
-                    }
-                }
-                "bracketMath" -> {
-                    val close = source.indexOf("\\]", nextIdx + 2).takeIf { it in 0 until end }
-                    val parsed = close?.let { mathNode(nextIdx + 2, it, it + 2) }
-                    if (parsed != null) {
-                        nodes.add(parsed)
-                        cursor = close + 2
-                    } else {
-                        nodes.add(InlineNode.Text(source.substring(nextIdx, end)))
-                        break
-                    }
-                }
-                "doubleDollarMath" -> {
-                    val close = source.indexOf("$$", nextIdx + 2).takeIf { it in 0 until end }
-                    val parsed = close?.let { mathNode(nextIdx + 2, it, it + 2) }
-                    if (parsed != null) {
-                        nodes.add(parsed)
-                        cursor = close + 2
-                    } else {
-                        nodes.add(InlineNode.Text("$$"))
-                        cursor = nextIdx + 2
-                    }
-                }
-                "dollarMath" -> {
-                    val close = findInlineDollarClose(source, nextIdx + 1).takeIf { it in 0 until end }
-                    val latex = close?.let { source.substring(nextIdx + 1, it).trim() }
-                    if (close != null && latex != null && looksLikeMath(latex, source.getOrNull(nextIdx - 1), source.getOrNull(close + 1))) {
-                        nodes.add(mathNode(nextIdx + 1, close, close + 1)!!)
-                        cursor = close + 1
-                    } else {
-                        nodes.add(InlineNode.Text("$"))
-                        cursor = nextIdx + 1
-                    }
-                }
-                "bold" -> {
-                    val close = source.indexOf("**", nextIdx + 2).takeIf { it in 0 until end }
-                    if (close != null) {
-                        nodes.add(InlineNode.Span(SpanStyle(fontWeight = FontWeight.Bold), parseRange(nextIdx + 2, close)))
-                        cursor = close + 2
-                    } else {
-                        nodes.add(InlineNode.Text("**"))
-                        cursor = nextIdx + 2
-                    }
-                }
-                "strike" -> {
-                    val close = source.indexOf("~~", nextIdx + 2).takeIf { it in 0 until end }
-                    if (close != null) {
-                        nodes.add(InlineNode.Span(SpanStyle(textDecoration = TextDecoration.LineThrough), parseRange(nextIdx + 2, close)))
-                        cursor = close + 2
-                    } else {
-                        nodes.add(InlineNode.Text("~~"))
-                        cursor = nextIdx + 2
-                    }
-                }
-                "italic" -> {
-                    val close = source.indexOf("*", nextIdx + 1).takeIf { it in 0 until end }
-                    if (close != null && close > nextIdx + 1) {
-                        nodes.add(InlineNode.Span(SpanStyle(fontStyle = FontStyle.Italic), parseRange(nextIdx + 1, close)))
-                        cursor = close + 1
-                    } else {
-                        nodes.add(InlineNode.Text("*"))
-                        cursor = nextIdx + 1
-                    }
-                }
-                "code" -> {
-                    val close = source.indexOf("`", nextIdx + 1).takeIf { it in 0 until end }
-                    if (close != null) {
-                        nodes.add(InlineNode.Code(source.substring(nextIdx + 1, close)))
-                        cursor = close + 1
-                    } else {
-                        nodes.add(InlineNode.Text("`"))
-                        cursor = nextIdx + 1
-                    }
-                }
-                "link" -> {
-                    val labelEnd = source.indexOf("]", nextIdx + 1).takeIf { it in 0 until end }
-                    val parenOpen = labelEnd?.let { source.getOrNull(it + 1) }
-                    val parenClose = if (parenOpen == '(') source.indexOf(")", labelEnd + 2).takeIf { it in 0 until end } else null
-                    if (labelEnd != null && parenClose != null) {
-                        val labelText = source.substring(nextIdx + 1, labelEnd)
-                        val url = source.substring(labelEnd + 2, parenClose)
-                        nodes.add(
-                            if (isCitationLabel(labelText)) {
-                                InlineNode.Citation(labelText, url)
-                            } else {
-                                InlineNode.Link(parseRange(nextIdx + 1, labelEnd), url)
-                            }
-                        )
-                        cursor = parenClose + 1
-                    } else {
-                        nodes.add(InlineNode.Text("["))
-                        cursor = nextIdx + 1
-                    }
-                }
-            }
-        }
-        return nodes
-    }
-
-    private fun mathNode(contentStart: Int, contentEnd: Int, rawEnd: Int): InlineNode.Math? {
-        val latex = source.substring(contentStart, contentEnd).trim()
-        if (latex.isEmpty()) return null
-        return InlineNode.Math(nextMathId++, latex, source.substring(contentStart - delimiterPrefixLength(contentStart), rawEnd))
-    }
-
-    private fun delimiterPrefixLength(contentStart: Int): Int =
-        when {
-            contentStart >= 2 && source.startsWith("\\(", contentStart - 2) -> 2
-            contentStart >= 2 && source.startsWith("\\[", contentStart - 2) -> 2
-            contentStart >= 2 && source.startsWith("$$", contentStart - 2) -> 2
-            else -> 1
-        }
-
-    private fun indexOfDoubleDollar(from: Int, end: Int): Int {
-        var i = source.indexOf("$$", from)
-        while (i != -1 && i < end) {
-            if (!isEscaped(source, i)) return i
-            i = source.indexOf("$$", i + 2)
-        }
-        return -1
-    }
-
-    private fun indexOfDollar(from: Int, end: Int): Int {
-        var i = source.indexOf("$", from)
-        while (i != -1 && i < end) {
-            if (!isEscaped(source, i) && source.getOrNull(i + 1) != '$' && source.getOrNull(i - 1) != '$') return i
-            i = source.indexOf("$", i + 1)
-        }
-        return -1
-    }
-}
-
-@Composable
-private fun InlineMarkdownText(
-    text: String,
-    color: Color,
-    linkColor: Color,
-    style: TextStyle,
-    modifier: Modifier = Modifier
-) {
-    val nodes = remember(text) { InlineParser(text).parse() }
-    val mathNodes = remember(nodes) { nodes.collectMathNodes() }
-    val hasMath = mathNodes.isNotEmpty()
-    if (!hasMath) {
-        val annotated = remember(text, linkColor) { parseInlineStyles(text, linkColor) }
-        Text(annotated, style = style.copy(color = color), modifier = modifier)
-        return
-    }
-
-    val textStyle = style.copy(color = color, lineHeight = lineHeightForInlineMath(style))
-    val config = latexConfigFor(style, scale = 0.86f)
-    val measurer = rememberLatexMeasurer(config)
-    val density = LocalDensity.current
-    val darkTheme = isSystemInDarkTheme()
-    val measured = remember(mathNodes, config, darkTheme) {
-        mathNodes.associate { math -> math.id to measurer.measure(math.latex, config, darkTheme) }
-    }
-    val annotated = remember(nodes, measured, linkColor) {
-        buildAnnotatedString {
-            appendNodes(nodes, linkColor, measured.keys)
-        }
-    }
-    val inlineContent = remember(mathNodes, measured, config, density, darkTheme) {
-        buildMap {
-            mathNodes.forEach { math ->
-                val dims = measured[math.id] ?: return@forEach
-                put(
-                    "math_${math.id}",
-                    InlineTextContent(
-                        placeholder = Placeholder(
-                            width = with(density) { dims.widthPx.toSp() },
-                            height = with(density) { dims.heightPx.toSp() },
-                            placeholderVerticalAlign = PlaceholderVerticalAlign.TextCenter
-                        )
-                    ) {
-                        Latex(
-                            latex = math.latex,
-                            config = config,
-                            isDarkTheme = darkTheme
-                        )
-                    }
-                )
-            }
-        }
-    }
-
-    Text(
-        text = annotated,
-        inlineContent = inlineContent,
-        style = textStyle,
-        modifier = modifier
-    )
-}
-
-private fun List<InlineNode>.collectMathNodes(): List<InlineNode.Math> {
-    val out = mutableListOf<InlineNode.Math>()
-    fun visit(nodes: List<InlineNode>) {
-        nodes.forEach { node ->
-            when (node) {
-                is InlineNode.Math -> out.add(node)
-                is InlineNode.Span -> visit(node.children)
-                is InlineNode.Link -> visit(node.label)
-                else -> Unit
-            }
-        }
-    }
-    visit(this)
-    return out
-}
-
-private fun AnnotatedString.Builder.appendNodes(
-    nodes: List<InlineNode>,
-    linkColor: Color,
-    measuredMathIds: Set<Int>
-) {
-    nodes.forEach { node ->
-        when (node) {
-            is InlineNode.Text -> append(node.text)
-            is InlineNode.Math -> {
-                if (node.id in measuredMathIds) {
-                    appendInlineContent("math_${node.id}", node.raw)
-                } else {
-                    append(node.raw)
-                }
-            }
-            is InlineNode.Span -> withStyle(node.style) {
-                appendNodes(node.children, linkColor, measuredMathIds)
-            }
-            is InlineNode.Code -> withStyle(
-                SpanStyle(
-                    fontFamily = FontFamily.Monospace,
-                    background = Color.Gray.copy(alpha = 0.18f),
-                    fontWeight = FontWeight.Medium
-                )
-            ) { append(node.text) }
-            is InlineNode.Link -> withLink(
-                LinkAnnotation.Url(
-                    node.url,
-                    TextLinkStyles(SpanStyle(color = linkColor, textDecoration = TextDecoration.Underline))
-                )
-            ) {
-                appendNodes(node.label, linkColor, measuredMathIds)
-            }
-            is InlineNode.Citation -> withLink(
-                LinkAnnotation.Url(
-                    node.url,
-                    TextLinkStyles(
-                        SpanStyle(
-                            color = linkColor,
-                            background = linkColor.copy(alpha = 0.14f),
-                            fontWeight = FontWeight.SemiBold,
-                            fontSize = 11.sp,
-                            baselineShift = BaselineShift.Superscript,
-                        )
-                    )
-                )
-            ) { append(" ${node.label} ") }
-        }
-    }
-}
-
-private fun lineHeightForInlineMath(style: TextStyle): TextUnit {
-    val baseSize = if (style.fontSize == TextUnit.Unspecified) 16.sp else style.fontSize
-    val current = if (style.lineHeight == TextUnit.Unspecified) baseSize * 1.5f else style.lineHeight
-    return current * 1.12f
-}
-
-@Composable
-private fun latexConfigFor(style: TextStyle, scale: Float): LatexConfig {
-    val baseSize = if (style.fontSize == TextUnit.Unspecified) 16.sp else style.fontSize
-    return LatexConfig(
-        fontSize = baseSize * scale,
-        theme = LatexTheme.material3(),
-        accessibilityEnabled = true
-    )
-}
-
-/**
- * Parses inline markdown — **bold**, *italic*, ~~strikethrough~~, `code`, and [text](url) links —
- * into a styled [AnnotatedString]. Nested spans (e.g. bold containing code) are handled recursively.
- */
-fun parseInlineStyles(text: String, linkColor: Color): AnnotatedString = buildAnnotatedString {
-    appendInline(text, linkColor)
-}
-
-private fun AnnotatedString.Builder.appendInline(text: String, linkColor: Color) {
-    var cursor = 0
-    val n = text.length
-    while (cursor < n) {
-        var nextIdx = -1
-        var kind = ""
-        fun consider(idx: Int, k: String) {
-            if (idx != -1 && (nextIdx == -1 || idx < nextIdx)) { nextIdx = idx; kind = k }
-        }
-        consider(text.indexOf("**", cursor), "bold")
-        consider(text.indexOf("~~", cursor), "strike")
-        consider(text.indexOf("`", cursor), "code")
-        consider(text.indexOf("[", cursor), "link")
-        consider(indexOfItalic(text, cursor), "italic")
-
-        if (nextIdx == -1) {
-            append(text.substring(cursor))
-            break
-        }
-        if (nextIdx > cursor) append(text.substring(cursor, nextIdx))
-
-        when (kind) {
-            "bold" -> {
-                val close = text.indexOf("**", nextIdx + 2)
-                if (close != -1) {
-                    withStyle(SpanStyle(fontWeight = FontWeight.Bold)) {
-                        appendInline(text.substring(nextIdx + 2, close), linkColor)
-                    }
-                    cursor = close + 2
-                } else { append("**"); cursor = nextIdx + 2 }
-            }
-            "strike" -> {
-                val close = text.indexOf("~~", nextIdx + 2)
-                if (close != -1) {
-                    withStyle(SpanStyle(textDecoration = TextDecoration.LineThrough)) {
-                        appendInline(text.substring(nextIdx + 2, close), linkColor)
-                    }
-                    cursor = close + 2
-                } else { append("~~"); cursor = nextIdx + 2 }
-            }
-            "code" -> {
-                val close = text.indexOf("`", nextIdx + 1)
-                if (close != -1) {
-                    withStyle(
-                        SpanStyle(
-                            fontFamily = FontFamily.Monospace,
-                            background = Color.Gray.copy(alpha = 0.18f),
-                            fontWeight = FontWeight.Medium
-                        )
-                    ) { append(text.substring(nextIdx + 1, close)) }
-                    cursor = close + 1
-                } else { append("`"); cursor = nextIdx + 1 }
-            }
-            "italic" -> {
-                val close = text.indexOf("*", nextIdx + 1)
-                if (close != -1 && close > nextIdx + 1) {
-                    withStyle(SpanStyle(fontStyle = FontStyle.Italic)) {
-                        appendInline(text.substring(nextIdx + 1, close), linkColor)
-                    }
-                    cursor = close + 1
-                } else { append("*"); cursor = nextIdx + 1 }
-            }
-            "link" -> {
-                val labelEnd = text.indexOf("]", nextIdx + 1)
-                val parenOpen = if (labelEnd != -1) text.getOrNull(labelEnd + 1) else null
-                if (labelEnd != -1 && parenOpen == '(') {
-                    val parenClose = text.indexOf(")", labelEnd + 2)
-                    if (parenClose != -1) {
-                        val label = text.substring(nextIdx + 1, labelEnd)
-                        val url = text.substring(labelEnd + 2, parenClose)
-                        if (isCitationLabel(label)) {
-                            // Numbered citation ([1](url)) — render as a tiny raised pill
-                            // in the accent color instead of a bare underlined "1".
-                            withLink(
-                                LinkAnnotation.Url(
-                                    url,
-                                    TextLinkStyles(
-                                        SpanStyle(
-                                            color = linkColor,
-                                            background = linkColor.copy(alpha = 0.14f),
-                                            fontWeight = FontWeight.SemiBold,
-                                            fontSize = 11.sp,
-                                            baselineShift = BaselineShift.Superscript,
-                                        )
-                                    )
-                                )
-                            ) { append(" $label ") }
-                        } else {
-                            withLink(
-                                LinkAnnotation.Url(
-                                    url,
-                                    TextLinkStyles(SpanStyle(color = linkColor, textDecoration = TextDecoration.Underline))
-                                )
-                            ) { appendInline(label, linkColor) }
-                        }
-                        cursor = parenClose + 1
-                    } else { append("["); cursor = nextIdx + 1 }
-                } else { append("["); cursor = nextIdx + 1 }
-            }
-        }
-    }
-}
-
-/** A link label that is just a small number — the `[1](url)` citation convention. */
-private fun isCitationLabel(label: String): Boolean =
-    label.length in 1..3 && label.all { it.isDigit() }
-
-/** Index of a standalone `*` italic marker (not part of `**`, and opening a non-space run). */
-private fun indexOfItalic(text: String, from: Int): Int {
-    var p = text.indexOf("*", from)
-    while (p != -1) {
-        val isDouble = text.getOrNull(p + 1) == '*' || (p > 0 && text[p - 1] == '*')
-        val opensWord = text.getOrNull(p + 1)?.isWhitespace() == false
-        if (!isDouble && opensWord) return p
-        p = text.indexOf("*", p + 1)
-    }
-    return -1
-}
-
-@Composable
-fun CodeBlockItem(code: String, language: String?) {
-    val context = LocalContext.current
-    val darkTheme = isSystemInDarkTheme()
-    // Real syntax highlighting via the Highlights engine, memoized per (code, language, theme) so
-    // finished blocks never recompute and only the actively-streaming block re-tokenizes.
-    val highlighted = remember(code, language, darkTheme) {
-        highlightCode(code, language, darkTheme)
-    }
-    Surface(
-        modifier = Modifier.fillMaxWidth().padding(vertical = 4.dp),
-        shape = RoundedCornerShape(16.dp),
-        color = MaterialTheme.colorScheme.surfaceContainerHighest,
-    ) {
-        Column {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(MaterialTheme.colorScheme.surfaceContainerHigh)
-                    .padding(start = 16.dp, end = 8.dp, top = 6.dp, bottom = 6.dp),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = (language ?: "code").uppercase(),
-                    style = MaterialTheme.typography.labelMedium.copy(
-                        fontFamily = FontFamily.Monospace,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                )
-                IconButton(
-                    onClick = {
-                        val clipboard = context.getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
-                        val clip = ClipData.newPlainText("Copied Code", code)
-                        clipboard.setPrimaryClip(clip)
-                        Toast.makeText(context, "Code copied into clipboard", Toast.LENGTH_SHORT).show()
-                    },
-                    modifier = Modifier.size(28.dp)
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.ContentCopy,
-                        contentDescription = "Copy code",
-                        tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                        modifier = Modifier.size(16.dp)
-                    )
-                }
-            }
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .horizontalScroll(rememberScrollState())
-                    .padding(12.dp)
-            ) {
-                SelectionContainer {
-                    Text(
-                        text = highlighted,
-                        style = MaterialTheme.typography.bodyMedium.copy(
-                            color = MaterialTheme.colorScheme.onSurface,
-                            fontFamily = FontFamily.Monospace,
-                            fontSize = 13.sp,
-                            lineHeight = 18.sp
-                        )
-                    )
-                }
-            }
-        }
-    }
-}
-
-/** Tokenizes [code] with the Highlights engine and maps the result to a styled [AnnotatedString]. */
-private fun highlightCode(code: String, language: String?, darkTheme: Boolean): AnnotatedString {
-    val syntaxLanguage = language?.let { runCatching { SyntaxLanguage.getByName(it) }.getOrNull() }
-    val highlights = runCatching {
-        Highlights.Builder()
-            .code(code)
-            .theme(SyntaxThemes.default(darkMode = darkTheme))
-            .let { if (syntaxLanguage != null) it.language(syntaxLanguage) else it }
-            .build()
-            .getHighlights()
-    }.getOrNull().orEmpty()
-
-    return buildAnnotatedString {
-        append(code)
-        highlights.forEach { h ->
-            val start = h.location.start.coerceIn(0, code.length)
-            val end = h.location.end.coerceIn(start, code.length)
-            if (start == end) return@forEach
-            when (h) {
-                is ColorHighlight -> addStyle(SpanStyle(color = Color(h.rgb).copy(alpha = 1f)), start, end)
-                is BoldHighlight -> addStyle(SpanStyle(fontWeight = FontWeight.Bold), start, end)
-            }
         }
     }
 }
